qrcode_gen/requirements.txt
tk
pandas
qrcode
pillow
xlrd
openpyxl
pyshorteners